+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## 0.3.0 (2017-12-18)
+- Automatically convert namespace/separators between servers
+- Attempt to reconnect if pipe breaks
+- Fetch headers when checking for duplicates
+
 ## 0.2.0 (2017-11-18)
 - If unable to connect to account, will skip that account and log the error rather than exiting the program.